pyproject.toml
[project]
name = "chatrag"
version = "0.1.0"
description = "ChatBot with Streamlit and Langchain using RAG and Memory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.1",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
]
